(#6, #7) Started on parse() & reworking Arguments
The first steps of the parse() function are implemented, but before work can continue, we have to update the Arguments class first to get it up-to-date. Specifically, we are now reworking it to now store Arguments (or derivatives of them) anymore, but instead we use a custom ParsedArgument class, which is non-polymorphic.