(#6) (Almost) finished usage(), working on help()
We're a long way generating the usage string, except that we have to double-check the brackets of MultiGroups. Also, we changed the usage functions of individual arguments to not return a string but to work instead directly on a common stringstream. Otherwise, help strings are starting to get along as well.