(#6) Alpha release ready
Finally we have implemented enough so that we can compile it. Finished the help & usage generators, added additional modes for Options to be parsed (-otest, --o=test), allowed option values to be omitted if a default value is present, fixed a shitton of bugs and did other stuff. The basic parsing of flags and options seems to work - so on to Positionals and more obscure features.