(#6) Implementing Optional Variadic
Option parsing nearly seems finished, and already supports parsing variadic values. However, it now seems that the first non-value token after the last one is thrown away, which needs to be fixed - to this end, we are creating a peekv() function that should migrate the checking of enough values to the Tokenizer instead of each parser individually, and which also leaves the stream unaltered if not enough values could be read.