(#18) Re-implemented traversal & added ADLValue
While finishing up the ADLDirective class, we both re-implemented the traversal functions to be a) more usable and consistent, and b) properly allow the user to choose when to recurse and to return new nodes which override old ones. Additionally, already implemented skeleton for ADLValue s.t. ADLDirective doesn't complain as much (though it still does :( )