(#18) Parsing directives is done
The basic functionality seems to be in place - a very basic parses that should be able to parse directives. However, while working on it, it seems apparent that we need to overhaul the debug information passed with each token and therefore node - in particular, we also want node stops s.t. we can implement the nice, multi-character error colours and we need to pass the raw line to a Node. See issue (#23).