(#18, #27) Almost added Positionals
The only thing missing from a successful Positional parser is parsing the three dots to indicate it's variadic. However, currently, the parser has trouble disambiguating when to promote a type terminal to a Types non-terminal; in particular, it leaves the type as it is when the three dots are added, meaning that it can't parse the Positional as such. Will probably have to tweak the grammar slightly to make this distinction more clear