(#31) Change reference parsing to token
Instead of re-defining the way the ADL looks, we changed parsing by making a reference an official token. This still add minor ambiguities for the triple dot token, but these are solved with a little bit of tokenizer magic. We only need to tweak the parser a little bit to accept the new reference tokens.