(#31) We finally added the errors
For now, error handling in the basic parser is done. There are a lot of cases which could use specific errors, and when a raw_line is not passed on in its entirety, we may get weird results. Nonetheless, the basic functionality is there, and so we can move on to more exciting things to do.